+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the z/OS FTP Plug-in for Zowe CLI will be documented in this file.
 
+## Recent Changes
+
+- Added --rdw to `download dataset` command to download variable-length dataset.
+
 ## `1.5.0`
 
 - Expose meta data for Zowe Explorer FTP extension.